+++ plugin-hrm-form/src/components/Messaging/MessageItem/index.ts
@@ -1,0 +1,2 @@
+export { default as MessageItem } from './MessageItem';
+export type { GroupedMessage } from './MessageItem';
