+++ plugin-hrm-form/src/components/Messaging/MessageList/index.ts
@@ -1,0 +1,1 @@
+export { default as MessageList } from './MessageList';
